Fix: Celery app auto discovery not finding task
